Move `TestCompareGoogleapis` to unix-only
